backend/cloudfunctions/bpm_detection/requirements.txt
google-cloud-firestore==2.0.2
google-cloud-storage==1.35.0
aubio==0.4.9
pydub==0.24.1
numpy==1.19.5

